requirements.txt
peewee
python-telegram-bot
geocoder